Refact SCRUM-184: 찜 목록 페이지, 모임 상세 페이지에 무한 스크롤 적용할 수 있도록 코드 셋팅
